+++ app/src/main/java/by/slizh/bankingapp/modelTest/Transaction.kt
@@ -1,4 +1,4 @@
-package by.slizh.bankingapp.model
+package by.slizh.bankingapp.modelTest
 
 data class Transaction(
     val company: String,
