+++ app/src/main/java/by/slizh/bankingapp/presentation/components/TransactionListItem.kt
@@ -1,4 +1,4 @@
-package by.slizh.bankingapp.components
+package by.slizh.bankingapp.presentation.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
@@ -17,7 +17,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import by.slizh.bankingapp.R
-import by.slizh.bankingapp.model.Transaction
+import by.slizh.bankingapp.modelTest.Transaction
 import by.slizh.bankingapp.ui.theme.Green
 import by.slizh.bankingapp.ui.theme.LightGrey
